Centralize event stream wrappers
