+++ packages/aws-event-stream/tests/unit/_private/test_serializers.py
@@ -7,10 +7,8 @@ from smithy_core.serializers import SerializeableShape
 from smithy_core.aio.types import AsyncBytesProvider
 from smithy_json import JSONCodec
 
-from aws_event_stream._private.serializers import (
-    EventSerializer,
-    AWSAsyncEventPublisher,
-)
+from aws_event_stream.aio import AWSEventPublisher
+from aws_event_stream._private.serializers import EventSerializer
 from aws_event_stream.events import EventMessage
 
 from . import EVENT_STREAM_SERDE_CASES, INITIAL_REQUEST_CASE, INITIAL_RESPONSE_CASE
@@ -46,7 +44,7 @@ def test_serialize_initial_response():
 
 async def test_publisher_closes_reader():
     writer = AsyncBytesProvider()
-    publisher: AWSAsyncEventPublisher[Any] = AWSAsyncEventPublisher(
+    publisher: AWSEventPublisher[Any] = AWSEventPublisher(
         payload_codec=JSONCodec(), async_writer=writer
     )
 
@@ -59,7 +57,7 @@ async def test_publisher_closes_reader():
 
 async def test_send_after_close():
     writer = AsyncBytesProvider()
-    publisher: AWSAsyncEventPublisher[Any] = AWSAsyncEventPublisher(
+    publisher: AWSEventPublisher[Any] = AWSEventPublisher(
         payload_codec=JSONCodec(), async_writer=writer
     )
 
@@ -71,7 +69,7 @@ async def test_send_after_close():
 
 async def test_send_to_closed_writer():
     writer = AsyncBytesProvider()
-    publisher: AWSAsyncEventPublisher[Any] = AWSAsyncEventPublisher(
+    publisher: AWSEventPublisher[Any] = AWSEventPublisher(
         payload_codec=JSONCodec(), async_writer=writer
     )
