+++ packages/smithy-core/src/smithy_core/aio/eventstream.py
@@ -1,75 +1,19 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
-from typing import Any, Protocol, Self
+from asyncio import Future
+from typing import Any, Self
 
-from smithy_core.deserializers import DeserializeableShape
-from smithy_core.serializers import SerializeableShape
+from ..deserializers import DeserializeableShape
+from ..serializers import SerializeableShape
 
-
-class AsyncEventPublisher[E: SerializeableShape](Protocol):
-    """Asynchronously sends events to a service.
-
-    This may be used as a context manager to ensure the stream is closed before exiting.
-    """
-
-    async def send(self, event: E) -> None:
-        """Sends an event to the service.
-
-        :param event: The event to send.
-        """
-        ...
-
-    async def close(self) -> None:
-        """Closes the event stream."""
-        ...
-
-    async def __aenter__(self) -> Self:
-        return self
-
-    async def __aexit__(self, exc_type: Any, exc_value: Any, traceback: Any):
-        await self.close()
+from .interfaces.eventstream import EventPublisher, EventReceiver
 
 
-class AsyncEventReceiver[E: DeserializeableShape](Protocol):
-    """Asynchronously receives events from a service.
-
-    Events may be received via the ``receive`` method or by using this class as
-    an async iterable.
-
-    This may also be used as a context manager to ensure the stream is closed before
-    exiting.
-    """
-
-    async def receive(self) -> E | None:
-        """Receive a single event from the service.
-
-        :returns: An event or None. None indicates that no more events will be sent by
-            the service.
-        """
-        ...
-
-    async def close(self) -> None:
-        """Closes the event stream."""
-        ...
-
-    async def __anext__(self) -> E:
-        result = await self.receive()
-        if result is None:
-            await self.close()
-            raise StopAsyncIteration
-        return result
-
-    def __aiter__(self) -> Self:
-        return self
-
-    async def __enter__(self) -> Self:
-        return self
-
-    async def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any):
-        await self.close()
-
-
-class DuplexEventStream[I: SerializeableShape, O: DeserializeableShape, R](Protocol):
+class DuplexEventStream[
+    IE: SerializeableShape,
+    OE: DeserializeableShape,
+    O: DeserializeableShape,
+]:
     """An event stream that both sends and receives messages.
 
     To ensure that streams are closed upon exiting, this class may be used as an async
@@ -106,46 +50,36 @@ class DuplexEventStream[I: SerializeableShape, O: DeserializeableShape, R](Proto
                             return
     """
 
-    input_stream: AsyncEventPublisher[I]
+    input_stream: EventPublisher[IE]
     """An event stream that sends events to the service."""
 
-    # Exposing response and output_stream via @property allows implementations that
-    # don't have it immediately available to do things like put a future in
-    # await_output or otherwise reasonably implement that method while still allowing
-    # them to inherit directly from the protocol class.
-    _output_stream: AsyncEventReceiver[O] | None = None
-    _response: R | None = None
+    output_stream: EventReceiver[OE] | None = None
+    """An event stream that receives events from the service.
 
-    @property
-    def output_stream(self) -> AsyncEventReceiver[O] | None:
-        """An event stream that receives events from the service.
+    This value may be None until ``await_output`` has been called.
 
-        This value may be None until ``await_output`` has been called.
+    This value will also be None if the operation has no output stream.
+    """
 
-        This value will also be None if the operation has no output stream.
-        """
-        return self._output_stream
+    response: O | None = None
+    """The initial response from the service.
 
-    @output_stream.setter
-    def output_stream(self, value: AsyncEventReceiver[O]) -> None:
-        self._output_stream = value
+    This value may be None until ``await_output`` has been called.
 
-    @property
-    def response(self) -> R | None:
-        """The initial response from the service.
+    This may include context necessary to interpret output events or prepare
+    input events. It will always be available before any events.
+    """
 
-        This value may be None until ``await_output`` has been called.
+    def __init__(
+        self,
+        *,
+        input_stream: EventPublisher[IE],
+        output_future: Future[tuple[O, EventReceiver[OE]]],
+    ) -> None:
+        self.input_stream = input_stream
+        self._output_future = output_future
 
-        This may include context necessary to interpret output events or prepare
-        input events. It will always be available before any events.
-        """
-        return self._response
-
-    @response.setter
-    def response(self, value: R) -> None:
-        self._response = value
-
-    async def await_output(self) -> tuple[R, AsyncEventReceiver[O]]:
+    async def await_output(self) -> tuple[O, EventReceiver[OE]]:
         """Await the operation's output.
 
         The EventStream will be returned as soon as the input stream is ready to
@@ -164,7 +98,8 @@ class DuplexEventStream[I: SerializeableShape, O: DeserializeableShape, R](Proto
         :returns: A tuple containing the initial response and output stream. If the
             operation has no output stream, the second value will be None.
         """
-        ...
+        self.response, self.output_stream = await self._output_future
+        return self.response, self.output_stream
 
     async def close(self) -> None:
         """Closes the event stream.
@@ -184,7 +119,7 @@ class DuplexEventStream[I: SerializeableShape, O: DeserializeableShape, R](Proto
         await self.close()
 
 
-class InputEventStream[I: SerializeableShape, R](Protocol):
+class InputEventStream[IE: SerializeableShape, O]:
     """An event stream that streams messages to the service.
 
     To ensure that streams are closed upon exiting, this class may be used as an async
@@ -203,47 +138,41 @@ class InputEventStream[I: SerializeableShape, R](Protocol):
                 await stream.await_output()
     """
 
-    input_stream: AsyncEventPublisher[I]
+    input_stream: EventPublisher[IE]
     """An event stream that sends events to the service."""
 
-    # Exposing response via @property allows implementations that don't have it
-    # immediately available to do things like put a future in await_output or
-    # otherwise reasonably implement that method while still allowing them to
-    # inherit directly from the protocol class.
-    _response: R | None = None
+    response: O | None = None
+    """The initial response from the service.
 
-    @property
-    def response(self) -> R | None:
-        """The initial response from the service.
+    This value may be None until ``await_output`` has been called.
 
-        This value may be None until ``await_output`` has been called.
+    This may include context necessary to interpret output events or prepare
+    input events. It will always be available before any events.
+    """
 
-        This may include context necessary to interpret output events or prepare
-        input events. It will always be available before any events.
+    def __init__(
+        self,
+        *,
+        input_stream: EventPublisher[IE],
+        output_future: Future[O],
+    ) -> None:
+        self.input_stream = input_stream
+        self._output_future = output_future
+
+    async def await_output(self) -> O:
+        """Await the operation's initial response.
+
+        The EventStream will be returned as soon as the input stream is ready to receive
+        events, which may be before the initial response has been received and the
+        service is ready to send events.
+
+        Awaiting this method will wait until the initial response was received.
+
+        :returns: The service's initial response.
         """
-        return self._response
-
-    @response.setter
-    def response(self, value: R) -> None:
-        self._response = value
-
-    async def await_output(self) -> R:
-        """Await the operation's output.
-
-        The InputEventStream will be returned as soon as the input stream is ready to
-        receive events, which may be before the initial response has been received.
-
-        Awaiting this method will wait until the initial response was received. The
-        operation response will be returned by this operation and also cached in
-        ``response``.
-
-        The default implementation of this method performs the caching behavior,
-        delegating to the abstract ``_await_output`` method to actually retrieve the
-        operation response.
-
-        :returns: The operation's response.
-        """
-        ...
+        if self.response is None:
+            self.response = await self._output_future
+        return self.response
 
     async def close(self) -> None:
         """Closes the event stream."""
@@ -256,7 +185,7 @@ class InputEventStream[I: SerializeableShape, R](Protocol):
         await self.close()
 
 
-class OutputEventStream[O: DeserializeableShape, R](Protocol):
+class OutputEventStream[OE: DeserializeableShape, O: DeserializeableShape]:
     """An event stream that streams messages from the service.
 
     To ensure that streams are closed upon exiting, this class may be used as an async
@@ -277,18 +206,22 @@ class OutputEventStream[O: DeserializeableShape, R](Protocol):
                             return
     """
 
-    output_stream: AsyncEventReceiver[O]
+    output_stream: EventReceiver[OE]
     """An event stream that receives events from the service.
 
     This value will also be None if the operation has no output stream.
     """
 
-    response: R
+    response: O
     """The initial response from the service.
 
     This may include context necessary to interpret output events or prepare input
     events. It will always be available before any events.
     """
+
+    def __init__(self, output_stream: EventReceiver[OE], output: O) -> None:
+        self.output_stream = output_stream
+        self.response = output
 
     async def close(self) -> None:
         """Closes the event stream."""
