--- packages/smithy-event-stream/NOTICE
@@ -1,1 +1,0 @@
-Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
