--- packages/smithy-event-stream/pyproject.toml
@@ -1,21 +1,0 @@
-[project]
-name = "smithy-event-stream"
-version = "0.0.1"
-description = "Smithy event stream interfaces and core components."
-readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-    "smithy-core",
-]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build]
-exclude = [
-  "tests",
-]
-
-[tool.ruff]
-src = ["src"]
