--- packages/smithy-event-stream/src/smithy_event_stream/__init__.py
@@ -1,6 +1,0 @@
-# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-import importlib.metadata
-
-__version__: str = importlib.metadata.version("smithy-event-stream")
