--- packages/smithy-event-stream/src/smithy_event_stream/aio/__init__.py
@@ -1,2 +1,0 @@
-# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
